Refactor grading criterion handling: remove structured grading criterion model, update imports
