+++ athena/athena/models/db_exercise.py
@@ -18,8 +18,6 @@ class DBExercise(Model, Base):
     grading_criteria = Column(JSON, nullable=True)
     meta = Column(JSON, nullable=False)
 
-    structured_grading_criterion = relationship("DBStructuredGradingCriterion", back_populates="exercise", uselist=False)  # Use uselist=False for one-to-one
-
     # Polymorphism, discriminator attribute
     type = Column(SqlEnum(ExerciseType), index=True, nullable=False)
