--- athena/athena/schemas/structured_grading_criterion.py
@@ -1,7 +1,0 @@
-from typing import List
-from athena.schemas.grading_criterion import GradingCriterion
-from pydantic import BaseModel
-
-
-class StructuredGradingCriterion(BaseModel):
-    criteria: List[GradingCriterion]
