+++ modules/modeling/module_modeling_llm/module_modeling_llm/core/generate_suggestions.py
@@ -1,4 +1,4 @@
-from athena.schemas.structured_grading_criterion import StructuredGradingCriterion
+from athena.schemas.grading_criterion import StructuredGradingCriterion
 from langchain_core.output_parsers import PydanticOutputParser
 from langchain_core.prompts import ChatPromptTemplate
